dont delete brain

+++ code/modules/mob/living/carbon/alien/organs.dm
@@ -185,6 +185,8 @@
 	for (var/obj/item/organ/internal/organ as anything in owner.organs)
 		if (organ == src)
 			continue
+		if (isbrain(organ))
+			continue
 		organ.Remove(owner)
 		qdel(organ)
